requirements.txt
Django==5.0.7
psycopg2-binary==2.9.9
django-simple-history==3.7.0
python-dotenv==1.0.1
